test: move the join-request notification spec into mobile-groups-join
